scripts: dist handles UI

+++ scripts/dist.sh
@@ -39,9 +39,15 @@ for FILENAME in $(find ./dist -mindepth 1 -maxdepth 1 -type f); do
     CONSULNAME="consul${EXTENSION}"
 
     pushd ./dist
-    cp ${FILENAME} ${CONSULNAME}
-    zip ./pkg/${VERSION}_${PLATFORM}.zip ${CONSULNAME}
-    rm ${CONSULNAME}
+
+    if [ "${FILENAME}" = "ui.zip" ]; then
+        cp ${FILENAME} ./pkg/${VERSION}_web_ui.zip
+    else
+        cp ${FILENAME} ${CONSULNAME}
+        zip ./pkg/${VERSION}_${PLATFORM}.zip ${CONSULNAME}
+        rm ${CONSULNAME}
+    fi
+
     popd
 done
